Remove Magic Number
The magic number is changed because it is difficult to maintain/repair.
In 'Input_Replay_Game' function and the 'Initialize_Showed_Board' function,
the const keyword is used so that the value can not be changed later.
And constant 'START' and 'REPLAY' is same use. So We combine them.

resolved : #9